requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
langgraph>=0.2.0
langchain-core>=0.3.0
redis==5.0.1
